refactor(codec): replaced/deprecated potentially redundant structures
deprecation: `CodecOf` and `DocumentCodecOf`
deprecation: `ProjectionCodec`
deprecation: `Codecs {  }`
deprecation: Projection and DocumentProjection

BREAKING-CHANGE: removed FieldCodec.provideDelegate(...) replace with `from {}`
BREAKING-CHANGE: changed type parameters of Map.get(FieldCodec)
BREAKING-CHANGE: changed type parameters of FieldCodec.from(Map)

feat: added FieldCodec.from(() -> Map) as replacement for FieldCodec.provideDelegate
feat: added CodecClass as replacement for `CodecOf`
