+++ codec/src/main/kotlin/Errors.kt
@@ -1,0 +1,30 @@
+/*
+ *	Copyright 2023 cufy.org and meemer.com
+ *
+ *	Licensed under the Apache License, Version 2.0 (the "License");
+ *	you may not use this file except in compliance with the License.
+ *	You may obtain a copy of the License at
+ *
+ *	    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *	Unless required by applicable law or agreed to in writing, software
+ *	distributed under the License is distributed on an "AS IS" BASIS,
+ *	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *	See the License for the specific language governing permissions and
+ *	limitations under the License.
+ */
+package org.cufy.codec
+
+/**
+ * An exception thrown when a codec operation fails.
+ *
+ * @author LSafer
+ * @since 2.0.0
+ */
+@Suppress("serial")
+open class CodecException(
+    message: String? = null,
+    cause: Throwable? = null,
+    enableSuppression: Boolean = true,
+    writableStackTrace: Boolean = true,
+) : RuntimeException(message ?: cause?.message, cause, enableSuppression, writableStackTrace)
